Added appropriate cast and thus removed an ANSI C warning.
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@26846 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/gen_conf.c
@@ -1144,8 +1144,8 @@ rt_public char eif_gen_is_uniform (EIF_REFERENCE obj, char code)
 	if (gdp->size == 0)
 		return EIF_TRUE;
 
-	if (codemap [code])
-		utype = codemap [code];
+	if (codemap [(int) code])
+		utype = codemap [(int) code];
 	else
 		utype = gdp->typearr [0];
